Move store list wire document type into types module
